requirements.txt
databricks-sdk>=0.12.0
streamlit>=1.32.0
jira>=3.5.1
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=6.0.1
requests>=2.31.0 